Fix build with latest BoringSSL
amlogic appears to have an outdated fork of wpa_supplicant, that is
missing all the recent fixes and improvements. This is, of course,
unsustainable and unsupportable.

As a temporary workaround, merge
https://android-review.googlesource.com/c/platform/external/wpa_supplicant_8/+/2123558

This does not fix the root issue. b/270158205 tracks unforking this
repository.

This repository previously caused problems in b/201683809, which
similarly delayed updating a security-critical dependency in Android.
Before that, it was b/182246642, where it looks like we updated the
dependency for this vendor.

Test: presubmit
Bug: 270298286
Bug: 270158205
Change-Id: I486c2451948a82484f8567a307b2e2c7f99b0d7e