audio: Play the audio only channel no audio output [1/1]
PD#SWPL-112345

Problem:
when switch to audio only channels, we receive the mute cmd and send the
'has video = 1' to audio hal which could effect the av sync and cause no
audio output.

Solution:
In the single demux decoder case, when receive the mute or unmute, it
should not control the start or stop logic and send the other cmd to
audio hal. Theredore, delete the logic.

Verify:
soddy/ohm/simth

Change-Id: Ib81355af02f53f5e858da5dc4730922d148e7030
Signed-off-by: Guanfeng Ma <guanfeng.ma@amlogic.com>