Take on new libui dependency in camera code
The android.hardware.camera.common@1.0-helper library is being converted
to support the new AIDL graphics mapper HAL and requires adjustments
here to prepare for that change.

Mainly there are some imports and "using" statements that are being
removed from HandleImporter.h and the library is going to depend on
android.hardware.graphics.allocator-V2 instead of V1.

Test: m
Bug: 285605852

Change-Id: Ic3018fdcc31e93d86a1dc7f07730caf4929ca64e